+++ R/pp_cfg.R
@@ -45,6 +45,7 @@ Config <- R6Class("pp_cfg",
                                  suitrankdie_layoutLF = suitrankdieLayoutGrobLF,
                                  suitrankdie_layoutRF = suitrankdieLayoutGrobRF,
                                  pawn_layout = pawnLayoutGrob,
+                                 preview_layout = previewLayoutGrob,
                                  pyramid_layout = pyramidLayoutGrob,
                                  pyramid_top = pyramidTopGrob,
                                  default_fn)
@@ -121,12 +122,16 @@ Config <- R6Class("pp_cfg",
                 return(private$cache[[key]])
             }
             if (grepl("die_layout", piece_side)) {
-                die_width <- get_style_element("width", "die_face", private$cfg, DIE_WIDTH)
+                die_width <- self$get_pp_width("die_face")
                 return (4 * die_width)
             }
             if (grepl("pyramid_layout", piece_side)) {
-                pyramid_height <- get_style_element("height", "pyramid_face", private$cfg, PYRAMID_HEIGHTS[rank], rank=rank)
+                pyramid_height <- self$get_pp_height("pyramid_face", rank=rank)
                 return (pyramid_height)
+            }
+            if (piece_side == "preview_layout") {
+                tile_width <- self$get_pp_width("tile_face")
+                return (3 * tile_width)
             }
             piece <- get_piece(piece_side)
             default <- switch(piece, 
@@ -150,22 +155,26 @@ Config <- R6Class("pp_cfg",
                 return(private$cache[[key]])
             }
             if (grepl("die_layout", piece_side)) {
-                die_height <- get_style_element("height", "die_face", private$cfg, DIE_WIDTH)
+                die_height <- self$get_pp_height("die_face")
                 return (3 * die_height)
             }
             if (grepl("pyramid_layout", piece_side)) {
-                pyramid_height <- get_style_element("height", "pyramid_face", private$cfg, PYRAMID_HEIGHTS[rank], rank=rank)
-                pyramid_width <- get_style_element("width", "pyramid_face", private$cfg, PYRAMID_WIDTHS[rank], rank=rank)
+                pyramid_height <- self$get_pp_height("pyramid_face", rank=rank)
+                pyramid_width <- self$get_pp_width("pyramid_face", rank=rank)
                 pyramid_diagonal <- sqrt(pyramid_height^2 + (0.5*pyramid_width)^2)
                 return (2 * pyramid_diagonal)
             }
             if (grepl("pyramid_top", piece_side)) {
-                pyramid_width <- get_style_element("width", "pyramid_face", private$cfg, PYRAMID_WIDTHS[rank], rank=rank)
+                pyramid_width <- self$get_pp_width("pyramid_face", rank=rank)
                 return(pyramid_width)
             }
             if (grepl("pawn_layout", piece_side)) {
-                pawn_height <- get_style_element("height", "pawn_face", private$cfg, PAWN_HEIGHT)
+                pawn_height <- self$get_pp_height("pawn_face")
                 return((2.5 / (7/8)) * pawn_height)
+            }
+            if (piece_side == "preview_layout") {
+                tile_height <- self$get_pp_height("tile_face")
+                return (3 * tile_height)
             }
             width <- self$get_pp_width(piece_side, rank)
             piece <- get_piece(piece_side)